requirements.txt
# Thanos Orchestrator Dependencies
anthropic>=0.75.0      # Claude API client
httpx>=0.27.0          # Async HTTP client (Oura API)
asyncpg>=0.29.0        # PostgreSQL async driver (WorkOS/Neon)
tiktoken>=0.5.0        # Accurate token counting

# CLI Utilities
yaspin>=3.0.0          # Animated loading spinners for CLI
python-dotenv>=1.0.0   # Load environment variables from .env
litellm>=1.0.0         # Unified LLM client

# MemOS - Hybrid Memory System
neo4j>=5.0.0           # Neo4j graph database driver (AuraDB)
chromadb>=0.5.0        # ChromaDB vector database
openai>=1.0.0          # OpenAI embeddings for vector search

# Google Calendar Integration
google-auth>=2.0.0               # Google authentication library
google-auth-oauthlib>=1.0.0      # OAuth 2.0 flow for Google APIs
google-api-python-client>=2.0.0  # Google Calendar API client

# Future: MCP SDK for full protocol support
# mcp>=1.0.0
